Move tests to new module
